+++ lsc/models/__init__.py
@@ -1,6 +1,7 @@
+from .arima_model import ARIMAModel
 from .base import Model, StateEstimate
 from .hmm import HMMModel
 from .kalman import KalmanModel
 from .switching import SwitchingModel
 
-__all__ = ["Model", "StateEstimate", "KalmanModel", "HMMModel", "SwitchingModel"]
+__all__ = ["Model", "StateEstimate", "KalmanModel", "HMMModel", "SwitchingModel", "ARIMAModel"]